+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 swiftmailer extension Change Log
 2.0.6 under development
 -----------------------
 
+- Enh #6: Added ability to specify custom mail header at `yii\swiftmailer\Message` (klimov-paul)
 - Enh #27: Added ability to specify message signature (klimov-paul)